Isomorphic and file uploads
This update includes major changes to directory structure. There are a
few left over files as a sort of paper trail that will be cleaned up
later. See server.js, main.jsx, and router.js for the main work in the
update. Isomorphic react and file uploads are implemented in this commit